WIP message refactor
