+++ pymongo/asynchronous/bulk.py
@@ -37,12 +37,8 @@ from bson.raw_bson import RawBSONDocument
 from pymongo import _csot, common
 from pymongo.asynchronous.client_session import AsyncClientSession, _validate_session_write_concern
 from pymongo.asynchronous.message import (
-    _DELETE,
-    _INSERT,
-    _UPDATE,
     _BulkWriteContext,
     _EncryptedBulkWriteContext,
-    _randint,
 )
 from pymongo.common import (
     validate_is_document_type,
@@ -56,6 +52,12 @@ from pymongo.errors import (
     OperationFailure,
 )
 from pymongo.helpers_shared import _RETRYABLE_ERROR_CODES, _get_wce_doc
+from pymongo.message import (
+    _DELETE,
+    _INSERT,
+    _UPDATE,
+    _randint,
+)
 from pymongo.read_preferences import ReadPreference
 from pymongo.write_concern import WriteConcern
