+++ pymongo/asynchronous/collection.py
@@ -41,8 +41,7 @@ from bson.objectid import ObjectId
 from bson.raw_bson import RawBSONDocument
 from bson.son import SON
 from bson.timestamp import Timestamp
-from pymongo import ASCENDING, _csot, common, helpers_shared
-from pymongo.asynchronous import message
+from pymongo import ASCENDING, _csot, common, helpers_shared, message
 from pymongo.asynchronous.aggregation import (
     _CollectionAggregationCommand,
     _CollectionRawAggregationCommand,
@@ -57,7 +56,6 @@ from pymongo.asynchronous.cursor import (
     AsyncCursor,
     AsyncRawBatchCursor,
 )
-from pymongo.asynchronous.message import _UNICODE_REPLACE_CODEC_OPTIONS
 from pymongo.collation import validate_collation_or_none
 from pymongo.common import _ecoc_coll_name, _esc_coll_name
 from pymongo.errors import (
@@ -67,6 +65,7 @@ from pymongo.errors import (
     OperationFailure,
 )
 from pymongo.helpers_shared import _check_write_command_response
+from pymongo.message import _UNICODE_REPLACE_CODEC_OPTIONS
 from pymongo.operations import (
     DeleteMany,
     DeleteOne,
