+++ pymongo/asynchronous/server.py
@@ -28,10 +28,10 @@ from typing import (
 
 from bson import _decode_all_selective
 from pymongo.asynchronous.helpers import _handle_reauth
-from pymongo.asynchronous.message import _convert_exception, _GetMore, _OpMsg, _Query
 from pymongo.errors import NotPrimaryError, OperationFailure
 from pymongo.helpers_shared import _check_command_response
 from pymongo.logger import _COMMAND_LOGGER, _CommandStatusMessage, _debug_log
+from pymongo.message import _convert_exception, _GetMore, _OpMsg, _Query
 from pymongo.response import PinnedResponse, Response
 
 if TYPE_CHECKING:
@@ -106,6 +106,34 @@ class Server:
         """Check the server's state soon."""
         self._monitor.request_check()
 
+    async def operation_to_command(
+        self, operation: Union[_Query, _GetMore], conn: AsyncConnection, apply_timeout: bool = False
+    ) -> tuple[dict[str, Any], str]:
+        is_query = isinstance(operation, _Query)
+        if is_query:
+            explain = "$explain" in operation.spec
+            cmd, db = operation.as_command()
+        else:
+            explain = False
+            cmd, db = operation.as_command(conn)
+        if operation.session:
+            await operation.session._apply_to(cmd, False, operation.read_preference, conn)
+            # Explain does not support readConcern.
+            if is_query and not explain and not operation.session.in_transaction:
+                operation.session._update_read_concern(cmd, conn)
+        # Support auto encryption
+        if operation.client._encrypter and not operation.client._encrypter._bypass_auto_encryption:
+            cmd = await operation.client._encrypter.encrypt(
+                operation.db, cmd, operation.codec_options
+            )
+
+        conn.add_server_api(cmd)
+        conn.send_cluster_time(cmd, operation.session, operation.client)
+        # Support CSOT
+        if apply_timeout:
+            conn.apply_timeout(operation.client, cmd=cmd if is_query else None)
+        return cmd, db
+
     @_handle_reauth
     async def run_operation(
         self,
@@ -122,26 +150,26 @@ class Server:
         cursors.
         Can raise ConnectionFailure, OperationFailure, etc.
 
-        :param conn: A AsyncConnection instance.
+        :param conn: An AsyncConnection instance.
         :param operation: A _Query or _GetMore object.
         :param read_preference: The read preference to use.
         :param listeners: Instance of _EventListeners or None.
         :param unpack_res: A callable that decodes the wire protocol response.
+        :param client: An AsyncMongoClient instance.
         """
-        duration = None
         assert listeners is not None
         publish = listeners.enabled_for_commands
         start = datetime.now()
 
         use_cmd = operation.use_command(conn)
         more_to_come = operation.conn_mgr and operation.conn_mgr.more_to_come
+        cmd, dbn = await self.operation_to_command(operation, conn, use_cmd)
         if more_to_come:
             request_id = 0
         else:
-            message = await operation.get_message(read_preference, conn, use_cmd)
+            message = operation.get_message(read_preference, conn, use_cmd)
             request_id, data, max_doc_size = self._split_message(message)
 
-        cmd, dbn = await operation.as_command(conn)
         if _COMMAND_LOGGER.isEnabledFor(logging.DEBUG):
             _debug_log(
                 _COMMAND_LOGGER,
@@ -160,7 +188,6 @@ class Server:
             )
 
         if publish:
-            cmd, dbn = await operation.as_command(conn)
             if "$db" not in cmd:
                 cmd["$db"] = dbn
             assert listeners is not None
