+++ test/test_change_stream.py
@@ -45,9 +45,9 @@ from pymongo.errors import (
     OperationFailure,
     ServerSelectionTimeoutError,
 )
+from pymongo.message import _CursorAddress
 from pymongo.read_concern import ReadConcern
 from pymongo.synchronous.command_cursor import CommandCursor
-from pymongo.synchronous.message import _CursorAddress
 from pymongo.write_concern import WriteConcern
