+++ test/test_pooling.py
@@ -24,10 +24,9 @@ import time
 
 from bson.codec_options import DEFAULT_CODEC_OPTIONS
 from bson.son import SON
-from pymongo import MongoClient, timeout
+from pymongo import MongoClient, message, timeout
 from pymongo.errors import AutoReconnect, ConnectionFailure, DuplicateKeyError
 from pymongo.hello_compat import HelloCompat
-from pymongo.synchronous import message
 
 sys.path[0:0] = [""]
