+++ test/test_read_preferences.py
@@ -39,6 +39,7 @@ from test.version import Version
 
 from bson.son import SON
 from pymongo.errors import ConfigurationError, OperationFailure
+from pymongo.message import _maybe_add_read_preference
 from pymongo.read_preferences import (
     MovingAverage,
     Nearest,
@@ -51,7 +52,6 @@ from pymongo.read_preferences import (
 from pymongo.server_description import ServerDescription
 from pymongo.server_selectors import Selection, readable_server_selector
 from pymongo.server_type import SERVER_TYPE
-from pymongo.synchronous.message import _maybe_add_read_preference
 from pymongo.synchronous.mongo_client import MongoClient
 from pymongo.write_concern import WriteConcern
